+++ src/personalized_catalog/scripts/personalized_catalog_allocation_alvi.py
@@ -327,7 +327,7 @@ def main() -> None:  # noqa: D103
     upper_store_banner = store_banner.upper()
     lower_store_banner = store_banner.lower()
 
-    monthid = pd.to_datetime(execution_date[:8] + '01').strftime('%Y%m').astype(int)
+    monthid = int(pd.to_datetime(execution_date[:8] + '01').strftime('%Y%m'))
 
     if store_banner == 'Alvi':
         organization_id = '08'
